HTMLheader {tools}	R Documentation

Generate a Standard HTML Header for R Help

Description

This function generates the standard HTML header used on R help pages.

Usage


HTMLheader(title = "R", logo = TRUE, up = NULL,
           top = file.path(Rhome, "doc/html/index.html"),
           Rhome = "",
           css = file.path(Rhome, "doc/html/R.css"),
           headerTitle = paste("R:", title),
           outputEncoding = "UTF-8")

Arguments

title
The title to display and use in the HTML headers. Should have had any HTML escaping already done.

logo
Whether to display the R logo after the title.

up
Which page (if any) to link to on the “up” button.

top
Which page (if any) to link to on the “top” button.

Rhome
A relative path to the R home directory. See the ‘Details’.

css
The relative URL for the Cascading Style Sheet.

headerTitle
The title used in the headers.

outputEncoding
The declared encoding for the whole page.

Details

The up and top links should be relative to the current page. The Rhome path default works with dynamic help; for static help, a relative path (e.g., ‘../..’) to it should be used.

Value

A character vector containing the lines of an HTML header which can be used to start a page in the R help system.

Examples


cat(HTMLheader("This is a sample header"), sep="\n")

[Package tools version 3.6.1 Index]