HTMLlinks {tools}	R Documentation

Collect HTML Links from Package Documentation

Description

Compute relative file paths for URLs to other package's installed HTML documentation.

Usage


findHTMLlinks(pkgDir = "", lib.loc = NULL, level = 0:2)

Arguments

pkgDir
the top-level directory of an installed package. The default indicates no package.

lib.loc
character vector describing the location of R library trees to scan: the default indicates .libPaths().

level
Which level(s) to include.

Details

findHTMLlinks tries to resolve links from one help page to another. It uses in decreasing priority

The package in pkgDir: this is used when converting HTML help for that package (level 0).

The base and recommended packages (level 1).

Other packages found in the library trees specified by lib.loc in the order of the trees and alphabetically within a library tree (level 2).

Value

A named character vector of file paths, relative to the ‘html’ directory of an installed package. So these are of the form ‘"../../somepkg/html/sometopic.html"’.

Author(s)

Duncan Murdoch, Brian Ripley

[Package tools version 3.6.1 Index]