Rcmd {tools}	R Documentation

R CMD Interface

Description

Invoke R CMD tools from within R.

Usage


Rcmd(args, ...)

Arguments

args
a character vector of arguments to R CMD.

...
arguments to be passed to system2.

Details

Provides a portable convenience interface to the R CMD mechanism by invoking the corresponding system commands (using the version of R currently used) via system2.

Value

See section “Value” in system2.

[Package tools version 3.6.1 Index]