Rd2txt_options {tools}	R Documentation

Set Formatting Options for Text Help

Description

This function sets various options for displaying text help.

Usage


Rd2txt_options(...)

Arguments

...
A list containing named options, or options passed as individual named arguments. See below for currently defined ones.

Details

This function persistently sets various formatting options for the Rd2txt function which is used in displaying text format help. Currently defined options are:

width

(default 80): The width of the output page.

minIndent

(default 10): The minimum indent to use in a list.

extraIndent

(default 4): The extra indent to use in each level of nested lists.

sectionIndent

(default 5): The indent level for a section.

sectionExtra

(default 2): The extra indentation for each nested section level.

itemBullet

(default "* ", with the asterisk replaced by a Unicode bullet in UTF-8 and most Windows locales): The symbol to use as a bullet in itemized lists.

enumFormat

: A function to format item numbers in enumerated lists.

showURLs

(default FALSE): Whether to show URLs when expanding \href tags.

code_quote

(default TRUE): Whether to render \code and similar with single quotes.

underline_titles

(default TRUE): Whether to render section titles with underlines (via backspacing).

Value

If called with no arguments, returns all option settings in a list. Otherwise, it changes the named settings and invisibly returns their previous values.

Author(s)

Duncan Murdoch

See Also

Rd2txt

Examples



# The itemBullet is locale-specific
saveOpts <- Rd2txt_options()
saveOpts
Rd2txt_options(minIndent = 4)
Rd2txt_options()
Rd2txt_options(saveOpts)
Rd2txt_options()


[Package tools version 3.6.1 Index]