add_datalist {tools}	R Documentation

Add a ‘datalist’ File to a Package

Description

The data() command with no arguments lists all the datasets available via data in attached packages, and to do so a per-package list is installed. Creating that list at install time can be slow for packages with huge datasets, and can be expedited by a supplying ‘data/datalist’ file.

Usage


add_datalist(pkgpath, force = FALSE)

Arguments

pkgpath
The path to a (source) package.

force
logical: can an existing ‘data/datalist’ file be over-written?

Details

R CMD build will call this function to add a data list to packages with 1MB or more of data.

It is also helpful to give a ‘data/datalist’ file in packages whose datasets have many dependencies, including loading the packages itself (and maybe others).

See Also

data.

The ‘Writing R Extensions’ manual.

[Package tools version 3.6.1 Index]