dependsOnPkgs {tools}	R Documentation

Find Reverse Dependencies

Description

Find ‘reverse’ dependencies of packages, that is those packages which depend on this one, and (optionally) so on recursively.

Usage


dependsOnPkgs(pkgs,
              dependencies = c("Depends", "Imports", "LinkingTo"),
              recursive = TRUE, lib.loc = NULL,
              installed =
              utils::installed.packages(lib.loc, fields = "Enhances"))

Arguments

pkgs
a character vector of package names.

dependencies
a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, and "most" currently shorthand for these apart from "Enhances".

recursive
logical: should reverse dependencies of reverse dependencies (and so on) be included?

lib.loc
a character vector of R library trees, or NULL for all known trees (see .libPaths).

installed
a result of calling installed.packages.

Value

A character vector of package names, which does not include any from pkgs.

Examples


## there are few dependencies in a vanilla R installation:
## lattice may not be installed
dependsOnPkgs("lattice")

[Package tools version 3.6.1 Index]