fileutils {tools}	R Documentation

File Utilities

Description

Utilities for testing and listing files, and manipulating file paths.

Usage

file_path_as_absolute(x)
file_path_sans_ext(x)
file_test(op, x, y)
list_files_with_exts(dir, exts, all.files = FALSE, full.names = TRUE)
list_files_with_type(dir, type, all.files = FALSE, full.names = TRUE)

Arguments

x,y 	 character vectors giving file paths.

op 	 a character string specifying the test to be performed. Unary tests (only x is used) are "-f" (existence and not being a directory) and "-d" (existence and directory); binary tests are "-nt" (newer than, using the modification dates) and "-ot".

dir 	 a character string with the path name to a directory.

exts 	 a character vector of possible file extensions.

all.files 	 a logical. If FALSE (default), only visible files are considered; if TRUE, all files are used.

full.names 	 a logical indicating whether the full paths of the files found are returned (default), or just the file names.

type 	 a character string giving the “type” of the files to be listed, as characterized by their extensions. Currently, possible values are "code" (R code), "data" (data sets), "demo" (demos), "docs" (R documentation), and "vignette" (vignettes).

Details

file_path_as_absolute turns a possibly relative file path absolute, performing tilde expansion if necessary. Currently, only a single existing path can be given.

file_path_sans_ext returns the file paths without extensions. (Only purely alphanumeric extensions are recognized.)

file_test performs shell-style file tests. Note that file.exists only tests for existence (test -e on some systems) but not for not being a directory.

list_files_with_exts returns the paths or names of the files in directory dir with extension matching one of the elements of exts. Note that by default, full paths are returned, and that only visible files are used.

list_files_with_type returns the paths of the files in dir of the given “type”, as determined by the extensions recognized by R. When listing R code and documentation files, files in OS-specific subdirectories are included if present. Note that by default, full paths are returned, and that only visible files are used.

See Also

file.path, file.info, list.files

Examples

dir <- file.path(R.home(), "library", "stats")
file_test("-d", dir)
file_test("-nt", file.path(dir, "R"), file.path(dir, "demo"))
list_files_with_exts(file.path(dir, "demo"), "R")
list_files_with_type(file.path(dir, "demo"), "demo") # the same
file_path_sans_ext(list.files(file.path(R.home(), "modules")))

[Package tools version 2.1.0 Index]