getVignetteInfo {tools}	R Documentation

Get Information on Installed Vignettes

Description

This function gets information on installed vignettes.

Usage


getVignetteInfo(package = NULL, lib.loc = NULL, all = TRUE)

Arguments

package
Which package to look in, or NULL for all packages.

lib.loc
Which library to look in.

all
Whether to search all installed packages, or just attached packages.

Value

A matrix with columns

Package
the name of the package

Dir
the directory where the package is installed

Topic
the name of the vignette

File
the base filename of the source of the vignette

Title
the title of the vignette

R
the tangled R source from the vignette

PDF
the PDF or HTML file for display

Note

The last column of the result is named PDF for historical reasons, but it may contain a filename of a PDF or HTML document.

See Also

pkgVignettes is a similar function that can work on an uninstalled package.

Examples



getVignetteInfo("grid")


[Package tools version 3.6.1 Index]