makeLazyLoading {tools}	R Documentation

Lazy Loading of Packages

Description

Tools for lazy loading of packages from a database.

Usage


makeLazyLoading(package, lib.loc = NULL, compress = TRUE,
                keep.source = getOption("keep.source.pkgs"),
                keep.parse.data = getOption("keep.parse.data.pkgs"),
                set.install.dir = NULL)

Arguments

package
package name string

lib.loc
library trees, as in library

keep.source
logical; should sources be kept when saving from source

keep.parse.data
logical; should parse data be kept with sources when keep.source is TRUE.

compress
logical; whether to compress entries on the database.

set.install.dir
character or null; if not null, set installation directory to this value before serializing internal data structures (for internal use, staged installation)

Details

A tool to set up packages for lazy loading from a database. For packages other than base you can use makeLazyLoading(package) to convert them to use lazy loading.

This is done when a package is installed.

Author(s)

Luke Tierney and Brian Ripley

[Package tools version 3.6.1 Index]