make_translations_pkg {tools}	R Documentation

Package the Current Translations in the R Sources

Description

A utility for R Core members to prepare a package of updated translations.

Usage


make_translations_pkg(srcdir, outDir = ".", append = "-1")

Arguments

srcdir
The R source directory.

outDir
The directory into which to place the prepared package.

append
The suffix for the package version number, e.g. 3.0.0-1 will be the default in R 3.0.0.

Details

This extracts the translations in a current R source distribution and packages them as a source package called translations which can be distributed on CRAN and installed by update.packages. This allows e.g. the translations shipped in R 3.x.y to be updated to those currently in R-patched, even by a user without administrative privileges.

The package has a Depends field which restricts it to versions 3.x.* for a single x.

[Package tools version 3.6.1 Index]