md5sum {tools}	R Documentation

Compute MD5 Checksums

Description

Compute the 32-byte MD5 hashes of one or more files.

Usage


md5sum(files)

Arguments

files
character. The paths of file(s) whose contents are to be hashed.

Details

A MD5 ‘hash’ or ‘checksum’ or ‘message digest’ is a 128-bit summary of the file contents represented by 32 hexadecimal digits. Files with different MD5 sums are different: only very exceptionally (and usually with the intent to deceive) are those with the same sums different.

On Windows all files are read in binary mode (as the md5sum utilities there do): on other OSes the files are read in the default mode (almost always text mode where there is more than one).

MD5 sums are used as a check that R packages have been unpacked correctly and not subsequently modified.

Value

A character vector of the same length as files, with names equal to files (possibly expanded). The elements will be NA for non-existent or unreadable files, otherwise a 32-character string of hexadecimal digits.

Source

The underlying C code was written by Ulrich Drepper and extracted from a 2001 release of glibc.

See Also

checkMD5sums

Examples


as.vector(md5sum(dir(R.home(), pattern = "^COPY", full.names = TRUE)))

[Package tools version 3.6.1 Index]