.print.via.format {tools}	R Documentation

Printing Utilities

Description

.print.via.format is a “prototype” print() method, useful, at least as a start, by a simple

    print.<myS3class>  <-  .print.via.format

Usage


.print.via.format(x, ...)

Arguments

x
object to be printed.

...
optional further arguments, passed to format.

Value

x, invisibly (by invisible()), as print methods should.

See Also

The print generic; its default method print.default (used for many basic implicit classes such as "numeric", "character" and arrays of them, lists etc).

Examples


## The function is simply defined as
 function (x, ...) {
    writeLines(format(x, ...))
    invisible(x)
 }

## is used for simple print methods in R, and as prototype for new methods.

[Package tools version 3.6.1 Index]