showNonASCII {tools}	R Documentation

Pick Out Non-ASCII Characters

Description

This function prints elements of a character vector which contain non-ASCII bytes, printing such bytes as a escape like <fc>.

Usage


showNonASCII(x)

showNonASCIIfile(file)

Arguments

x
a character vector.

file
path to a file.

Details

This was originally written to help detect non-portable text in files in packages.

It prints all element of x which contain non-ASCII characters, preceded by the element number and with non-ASCII bytes highlighted via iconv(sub = "byte").

Value

The elements of x containing non-ASCII characters will be returned invisibly.

Examples


out <- c(
"fa\xE7ile test of showNonASCII():",
"\\details{",
"   This is a good line",
"   This has an \xfcmlaut in it.",
"   OK again.",
"}")
f <- tempfile()
cat(out, file = f, sep = "\n")

showNonASCIIfile(f)
unlink(f)

[Package tools version 3.6.1 Index]