toTitleCase {tools}	R Documentation

Convert Titles to Title Case

Description

Convert a character vector to title case, especially package titles.

Usage


toTitleCase(text)

Arguments

text
a character vector.

Details

This is intended for English text only.

No definition of‘title case’ is universally accepted: all agree that ‘principal’ words are capitalized and common words like ‘for’ are not, but not which words fall into each category.

Generally words in all capitals are left alone: this implementation knows about conventional mixed-case words such as ‘LaTeX’ and ‘OpenBUGS’ and a few technical terms which are not usually capitalized such as ‘jar’ and ‘xls’. However, unknown technical terms will be capitalized unless they are single words enclosed in single quotes: names of packages and libraries should be quoted in titles.

Value

A character vector of the same length as text, without names.

[Package tools version 3.6.1 Index]