tools-defunct {tools}	R Documentation

Defunct Objects in Package tools

Description

The functions or variables listed here are defunct, and may get removed as soon as of the next release.

Usage

filePathAsAbsolute(x)
filePathSansExt(x)
fileTest(op, x, y)
listFilesWithExts(dir, exts, all.files = FALSE, full.names = TRUE)
listFilesWithType(dir, type, all.files = FALSE, full.names = TRUE)

See Also

tools-deprecated, Deprecated, Defunct

[Package tools version 2.1.0 Index]