undoc {tools}	R Documentation

Find Undocumented Objects

Description

Finds the objects in a package which are undocumented, in the sense that they are visible to the user (or data objects or S4 classes provided by the package), but no documentation entry exists.

Usage

undoc(package, dir, lib.loc = NULL)

Arguments

package 	 a character string naming an installed package.

dir 	 a character string specifying the path to a package's root source directory. This must contain the subdirectory ‘man’ with R documentation sources (in Rd format), and at least one of the ‘R’ or ‘data’ subdirectories with R code or data objects, respectively.

lib.loc 	 a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. The specified library trees are used to to search for package.

Details

This function is useful for package maintainers mostly. In principle, all user level R objects should be documented; note however that the precise rules for documenting methods of generic functions are still under discussion.

Value

An object of class "undoc" which is a list of character vectors containing the names of the undocumented objects split according to documentation type. This representation is still experimental, and might change in future versions.
There is a print method for nicely displaying the information contained in such objects.

See Also

codoc, QC

Examples

undoc("tools")                  # Undocumented objects in 'tools'

[Package tools version 2.1.0 Index]