vignetteInfo {tools}	R Documentation

Basic Information about a Vignette

Description

Provide basic information including package and dependency of a vignette from its source file.

Usage


vignetteInfo(file)

Arguments

file
file name of the vignette.

Value

a list with components, each a possibly empty character:

file
the basename of the file.

title
the vignette title.

depends
the package dependencies.

keywords
keywords if provided.

engine
the vignette engine such as "Sweave", "knitr", etc.

Note

vignetteInfo(file)$depends is a substitute for the deprecated vignetteDepends() functionality.

See Also

package_dependencies

Examples

## This may not be installed, as it requires lattice
  gridEx <- system.file("doc", "grid.Rnw", package = "grid")
  vi <- vignetteInfo(gridEx)
  str(vi)

[Package tools version 3.6.1 Index]