write_PACKAGES {tools}	R Documentation

Generating a PACKAGES file

Description

Generating a ‘PACKAGES’ file for a repository of source or Windows binary packages.

Usage

write_PACKAGES(dir, fields,
               type = c("source", "mac.binary", "win.binary"),
               verbose = FALSE)

Arguments

dir 	 Character vector describing the location of the repository (directory including source or binary packages) to generate the ‘PACKAGES’ file from and write it to.

fields 	 Optional, the fields to be used in the ‘PACKAGES’ file. The default are those needed by available.packages: “Package”, “Bundle”, “Priority”, “Version”, “Depends”, “Suggests”, “Imports” and “Contains”.

type 	 Type of packages: currently source .tar.gz archives and Windows binary .zip packages are supported. Defaults to "win.binary" on Windows and to "source" otherwise.

verbose 	 logical. Should packages be listed as they are processed?

Details

type = "win.binary" uses unz connections to read all ‘DESCRIPTION’ files contained in the (zipped) binary packages for Windows in the given directory dir, and builds a ‘PACKAGES’ file from these information.

Value

Invisibly returns the number of packages described in the resulting ‘PACKAGES’ file. If 0, no packages were found and no ‘PACKAGES’ file has been written.

Note

Processing .tar.gz archives to extract the ‘DESCRIPTION’ files is quite slow.

This function can be useful on other OSes to prepare a repository to be accessed by Windows machines, so type = "winBinary" should work on all OSes.

Author(s)

Uwe Ligges and R-core.

See Also

See read.dcf and write.dcf for reading ‘DESCRIPTION’ files and writing the ‘PACKAGES’ file.

Examples

## Not run: 
write_PACKAGES("c:/myFolder/myRepository")  # on Windows
write_PACKAGES("/pub/RWin/bin/windows/contrib/2.1",
               type="win.binary")  # on Linux
## End(Not run)

[Package tools version 2.1.0 Index]