CRANtools {tools}	R Documentation

CRAN Package Repository Tools

Description

Tools for obtaining information about current packages in the CRAN package repository, and their check status.

Usage

CRAN_package_db()

CRAN_check_results(flavors = NULL)
CRAN_check_details(flavors = NULL)
CRAN_check_issues()
summarize_CRAN_check_status(packages,
                            results = NULL,
                            details = NULL,
                            issues = NULL)

Arguments

packages
a character vector of package names.

flavors
a character vector of CRAN check flavor names, or NULL (default), corresponding to all available flavors.

results
the return value of CRAN_check_results() (default), or a subset of this.

details
the return value of CRAN_check_details() (default), or a subset of this.

issues
the return value of CRAN_check_issues() (default), or a subset of this.

Details

CRAN_package_db() returns a character data frame with most ‘DESCRIPTION’ metadata for the current packages in the CRAN package repository, including in particular the Description and Maintainer information not provided by utils::available.packages().

CRAN_check_results() returns a data frame with the basic CRAN package check results including timings, with columns Package, Flavor and Status giving the package name, check flavor, and overall check status, respectively.

CRAN_check_details() returns a data frame inheriting from class "check_details" (which has useful print and format methods) with details on the check results, providing check name, status and output for every non-OK check (via columns Check, Status and Output, respectively). Packages with all-OK checks are indicated via a * Check wildcard name and OK Status.

CRAN_check_issues() returns a character frame with additional check issues (including the memory-access check results made available from https://www.stats.ox.ac.uk/pub/bdr/memtests/ and the without-long-double check results from https://www.stats.ox.ac.uk/pub/bdr/noLD/), as a character frame with variables Package, Version, kind (an identifier for the issue) and href (a URL with information on the issue).

CRAN_memtest_notes() is now deprecated, with its functionality integrated into that of CRAN_check_issues().

Value

See ‘Details’. Note that the results are collated on CRAN: currently this is done in a locale which sorts aAbB ....

Which CRAN?

The main functions access a CRAN mirror specified by the environment variable R_CRAN_WEB, defaulting to one specified in the ‘repositories’ file (see setRepositories): if that specifies @CRAN@ (the default) then https://CRAN.R-project.org is used. (Note that options("repos") is not consulted.)

Note that these access parts of CRAN under ‘web/contrib’ and ‘web/packages’ so if you have specified a mirror of just ‘src/contrib’ for installing packages you will need to set R_CRAN_WEB to point to a full mirror.

Examples

## No test: 
## This can be rather slow, especially with a non-local CRAN mirror
## and would fail (slowly) without Internet access in that case.

set.seed(11)  # but the packages chosen will change as soon as CRAN does.
pdb <- CRAN_package_db()
dim(pdb)

## [1] 18608    66

## DESCRIPTION fields included:
colnames(pdb)

##  [1] "Package"                 "Version"                 "Priority"                "Depends"                
##  [5] "Imports"                 "LinkingTo"               "Suggests"                "Enhances"               
##  [9] "License"                 "License_is_FOSS"         "License_restricts_use"   "OS_type"                
## [13] "Archs"                   "MD5sum"                  "NeedsCompilation"        "Additional_repositories"
## [17] "Author"                  "Authors@R"               "Biarch"                  "BugReports"             
## [21] "BuildKeepEmpty"          "BuildManual"             "BuildResaveData"         "BuildVignettes"         
## [25] "Built"                   "ByteCompile"             "Classification/ACM"      "Classification/ACM-2012"
## [29] "Classification/JEL"      "Classification/MSC"      "Classification/MSC-2010" "Collate"                
## [33] "Collate.unix"            "Collate.windows"         "Contact"                 "Copyright"              
## [37] "Date"                    "Description"             "Encoding"                "KeepSource"             
## [41] "Language"                "LazyData"                "LazyDataCompression"     "LazyLoad"               
## [45] "MailingList"             "Maintainer"              "Note"                    "Packaged"               
## [49] "RdMacros"                "StagedInstall"           "SysDataCompression"      "SystemRequirements"     
## [53] "Title"                   "Type"                    "URL"                     "UseLTO"                 
## [57] "VignetteBuilder"         "ZipData"                 "Published"               "Path"                   
## [61] "X-CRAN-Comment"          "Reverse depends"         "Reverse imports"         "Reverse linking to"     
## [65] "Reverse suggests"        "Reverse enhances"

## Summarize publication dates:
summary(as.Date(pdb$Published))

##         Min.      1st Qu.       Median         Mean      3rd Qu.         Max. 
## "2006-03-15" "2018-05-22" "2020-06-22" "2019-08-07" "2021-06-07" "2021-12-12"

## Summarize numbers of packages according to maintainer:
summary(lengths(split(pdb$Package, pdb$Maintainer)))

##    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
##    1.00    1.00    1.00    1.72    2.00   66.00

## Packages with 'LASSO' in their Description:
pdb$Package[grepl("LASSO", pdb$Description)]

##  [1] "BBSSL"             "CARlasso"          "CovSelHigh"        "crrp"              "expandFunctions"  
##  [6] "extlasso"          "fastcmprsk"        "FeaLect"           "FindIt"            "fsMTS"            
## [11] "gcdnet"            "glassoFast"        "glmnetSE"          "glmvsd"            "higlasso"         
## [16] "HTLR"              "ISBF"              "MetabolicSurv"     "midasml"           "milr"             
## [21] "miselect"          "MTE"               "MVB"               "ncpen"             "nets"             
## [26] "onlineVAR"         "pgee.mixed"        "polywog"           "predhy"            "rbridge"          
## [31] "rqPen"             "rrat"              "SentimentAnalysis" "slasso"            "SparseBiplots"    
## [36] "SSLASSO"           "stR"               "SurvHiDim"

results <- CRAN_check_results()
## Available variables:
names(results)

##  [1] "Flavor"     "Package"    "Version"    "Priority"   "Maintainer" "Status"     "Flags"      "T_install" 
##  [9] "T_check"    "T_total"

## Tabulate overall check status according to flavor:
with(results, table(Flavor, Status))

##                                    Status
## Flavor                                 OK  NOTE WARNING ERROR FAILURE
##   r-devel-linux-x86_64-debian-clang 13095  5281      87   100       0
##   r-devel-linux-x86_64-debian-gcc   12880  5215     378    78      26
##   r-devel-linux-x86_64-fedora-clang 10296  8144      63    49       0
##   r-devel-linux-x86_64-fedora-gcc   11448  7008      57    47       0
##   r-devel-windows-x86_64-new-TK     11776  6549     112    81       3
##   r-devel-windows-x86_64-new-UL     12924  5473      38    95       8
##   r-devel-windows-x86_64-old        13028  5401      30    81       4
##   r-oldrel-macos-x86_64             13610  4559     185   181       3
##   r-oldrel-windows-ix86+x86_64      15947  2402      38   138       3
##   r-patched-linux-x86_64            13163  5292      88    12       1
##   r-release-linux-x86_64            13188  5287      85    16       1
##   r-release-macos-arm64             10846  7615      26    88      10
##   r-release-macos-x86_64            10954  7337     183    99       0
##   r-release-windows-ix86+x86_64     12848  5597      24    81       4

details <- CRAN_check_details()
## Available variables:
names(details)

## [1] "Package"    "Version"    "Flavor"     "Check"      "Status"     "Output"     "Flags"      "Maintainer"

## Tabulate checks according to their status:
tab <- with(details, table(Check, Status))
## Inspect some installation problems:
bad <- subset(details,
              ((Check == "whether package can be installed") &
               (Status != "OK")))
## Show a random sample of up to 6
head(bad[sample(seq_len(NROW(bad)), NROW(bad)), ])

Package 	Version 	Flavor 	Check 	Status 	Output 	Flags 	Maintainer

238615 	specmine.datasets 	0.0.2 	r-release-windows-ix86+x86_64 	whether package can be installed 	WARNING 	Found the following significant warnings:

Warning: namespace 'xcms' is not available and has been replaced See 'd:/Rcompile/CRANpkg/local/4.1/specmine.datasets.Rcheck/00install.out' for details. | |Miguel Rocha mrocha@di.uminho.pt | |60376 |WSGeometry |1.2 |r-devel-linux-x86_64-fedora-clang |whether package can be installed |WARNING |Found the following significant warnings: maaipm.cpp:303:22: warning: result of '10-4' is -10; did you mean '1e-4'? [-Wxor-used-as-pow] maaipm.cpp:346:22: warning: result of '10-4' is -10; did you mean '1e-4'? [-Wxor-used-as-pow] maaipm.cpp:534:24: warning: result of '10-4' is -10; did you mean '1e-4'? [-Wxor-used-as-pow] maaipm.cpp:577:24: warning: result of '10-4' is -10; did you mean '1e-4'? [-Wxor-used-as-pow] umaaipm.cpp:234:24: warning: result of '10-3' is -9; did you mean '1e-3'? [-Wxor-used-as-pow] umaaipm.cpp:274:24: warning: result of '10-3' is -9; did you mean '1e-3'? [-Wxor-used-as-pow] umaaipm.cpp:476:24: warning: result of '10-3' is -9; did you mean '1e-3'? [-Wxor-used-as-pow] umaaipm.cpp:515:24: warning: result of '10-3' is -9; did you mean '1e-3'? [-Wxor-used-as-pow] See ‘/data/gannet/ripley/R/packages/tests-clang/WSGeometry.Rcheck/00install.out’ for details. | |Florian Heinemann florian.heinemann@uni-goettingen.de | |215683 |RcppMLPACK |1.0.10-7 |r-release-macos-x86_64 |whether package can be installed |WARNING |Found the following significant warnings: WARNING: shared objects/dynamic libraries with hard-coded temporary installation paths See ‘/Volumes/Builds/packages/high-sierra-x86_64/results/4.1/RcppMLPACK.Rcheck/00install.out’ for details. | |Qiang Kou qkou@umail.iu.edu | |194568 |Rcplex |0.3-5 |r-release-macos-arm64 |whether package can be installed |ERROR |Installation failed. See ‘/Volumes/Builds/packages/big-sur-arm64/results/4.1/Rcplex.Rcheck/00install.out’ for details. | |Stefan Theussl Stefan.Theussl@R-project.org | |135145 |RGtk2 |2.20.36.2 |r-devel-windows-x86_64-old |whether package can be installed |WARNING |Found the following significant warnings: atkClasses.c:1054:25: warning: assignment to 'void ()(AtkObject *, guint, void *)' {aka 'void ()(struct _AtkObject , unsigned int, void *)'} from incompatible pointer type 'void ()(AtkObject , guint, AtkObject *)' {aka 'void ()(struct _AtkObject , unsigned int, struct _AtkObject *)'} [-Wincompatible-pointer-types] atkClasses.c:1062:34: warning: assignment to 'void ()(AtkObject , void *)' {aka 'void ()(struct _AtkObject *, void *)'} from incompatible pointer type 'void ()(AtkObject *, AtkObject *)' {aka 'void ()(struct _AtkObject , struct _AtkObject *)'} [-Wincompatible-pointer-types] atkClasses.c:1401:51: warning: passing argument 2 of 'object_class->active_descendant_changed' from incompatible pointer type [-Wincompatible-pointer-types] atkConversion.c:97:36: warning: the address of 'names' will always evaluate as 'true' [-Waddress] atkConversion.c:136:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] cairoConversion.c:54:35: warning: the address of 'pathNames' will always evaluate as 'true' [-Waddress] cairoConversion.c:149:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] cairoConversion.c:171:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] cairoConversion.c:196:36: warning: the address of 'listNames' will always evaluate as 'true' [-Waddress] cairoConversion.c:228:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] cairoConversion.c:267:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:249:37: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:267:36: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:302:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:335:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:389:34: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:419:36: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:454:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:496:36: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkConversion.c:663:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gdkFuncs.c:6000:36: warning: passing argument 2 of 'gdk_window_get_user_data' from incompatible pointer type [-Wincompatible-pointer-types] gioClasses.c:1974:17: warning: assignment to 'gssize ()(GOutputStream , const void *, gsize, GCancellable *, GError *)' {aka 'long long int ()(struct _GOutputStream *, const void *, long long unsigned int, struct _GCancellable *, struct _GError *)'} from incompatible pointer type 'gssize ()(GOutputStream *, const guchar *, gsize, GCancellable *, GError *)' {aka 'long long int ()(struct _GOutputStream *, const unsigned char *, long long unsigned int, struct _GCancellable *, struct _GError *)'} [-Wincompatible-pointer-types] gioClasses.c:1990:20: warning: assignment to 'void ()(GOutputStream *, const void *, gsize, int, GCancellable *, void ()(GObject , GAsyncResult *, void *), void *)' {aka 'void ()(struct _GOutputStream , const void *, long long unsigned int, int, struct _GCancellable *, void ()(struct _GObject , struct _GAsyncResult *, void *), void *)'} from incompatible pointer type 'void ()(GOutputStream , const guchar *, gsize, int, GCancellable *, void ()(GObject , GAsyncResult *, void *), void *)' {aka 'void ()(struct _GOutputStream , const unsigned char *, long long unsigned int, int, struct _GCancellable *, void ()(struct _GObject , struct _GAsyncResult *, void *), void *)'} [-Wincompatible-pointer-types] gioConversion.c:16:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] glib.c:265:39: warning: the address of 'classes' will always evaluate as 'true' [-Waddress] gtkAccessors.c:1232:43: warning: the address of 'val' will always evaluate as 'true' [-Waddress] gtkClasses.c:8831:28: warning: assignment to 'gboolean ()(GtkNotebook , gint)' {aka 'int ()(struct _GtkNotebook , int)'} from incompatible pointer type 'void ()(GtkNotebook , gint)' {aka 'void ()(struct _GtkNotebook , int)'} [-Wincompatible-pointer-types] gtkClasses.c:11314:21: warning: assignment to 'gboolean ()(GtkScrolledWindow , GtkScrollType, gboolean)' {aka 'int ()(struct _GtkScrolledWindow , enum , int)'} from incompatible pointer type 'void ()(GtkScrolledWindow , GtkScrollType, gboolean)' {aka 'void ()(struct _GtkScrolledWindow *, enum , int)'} [-Wincompatible-pointer-types] gtkClasses.c:24141:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkClasses.c:24219:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkClasses.c:24276:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkClasses.c:24298:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkConversion.c:25:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkConversion.c:58:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkConversion.c:102:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkConversion.c:167:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkConversion.c:209:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkConversion.c:256:35: warning: the address of 'names' will always evaluate as 'true' [-Waddress] gtkFuncs.c:3904:100: warning: the address of 'requisition' will always evaluate as 'true' [-Waddress] gtkFuncs.c:6593:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:18182:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:18202:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:18222:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:18256:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27313:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27334:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27354:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27374:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27393:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27412:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27432:88: warning: the address of 'start' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27432:226: warning: the address of 'end' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27453:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27473:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27604:88: warning: the address of 'start' will always evaluate as 'true' [-Waddress] gtkFuncs.c:27604:226: warning: the address of 'end' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29152:100: warning: the address of 'match_start' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29152:256: warning: the address of 'match_end' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29178:100: warning: the address of 'match_start' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29178:256: warning: the address of 'match_end' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29894:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29916:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:29960:100: warning: the address of 'target_iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:32492:106: warning: the address of 'selection_data' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33140:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33162:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33201:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33222:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33300:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33357:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33379:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33625:100: warning: the address of 'filter_iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33645:98: warning: the address of 'child_iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33792:96: warning: the address of 'sort_iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33829:98: warning: the address of 'child_iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:33991:146: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:34406:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:34427:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:34448:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:34468:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:34488:86: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkFuncs.c:37680:100: warning: the address of 'requisition' will always evaluate as 'true' [-Waddress] gtkFuncs.c:37714:100: warning: the address of 'requisition' will always evaluate as 'true' [-Waddress] gtkFuncs.c:58643:100: warning: the address of 'requisition' will always evaluate as 'true' [-Waddress] gtkManuals.c:748:25: warning: the address of 'value' will always evaluate as 'true' [-Waddress] gtkManuals.c:1056:25: warning: the address of 'value' will always evaluate as 'true' [-Waddress] gtkManuals.c:1090:25: warning: the address of 'value' will always evaluate as 'true' [-Waddress] gtkManuals.c:1318:298: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] gtkManuals.c:1347:298: warning: the address of 'iter' will always evaluate as 'true' [-Waddress] pangoAccessors.c:658:43: warning: the address of 'val' will always evaluate as 'true' [-Waddress] pangoConversion.c:85:35: warning: the address of 'classes' will always evaluate as 'true' [-Waddress] See 'd:/Rcompile/CRANpkg/local/4.2/RGtk2.Rcheck/00install.out' for details. | |ORPHANED | |75107 |Rblpapi |0.3.12 |r-devel-linux-x86_64-fedora-gcc |whether package can be installed |NOTE |Found the following notes/warnings: Non-staged installation was used See ‘/data/gannet/ripley/R/packages/tests-devel/Rblpapi.Rcheck/00install.out’ for details. | |Dirk Eddelbuettel edd@debian.org |

issues <- CRAN_check_issues()
head(issues)

Package 	Version 	kind 	href

3 	ISwR 	2.0-8 	ATLAS 	https://www.stats.ox.ac.uk/pub/bdr/Rblas/ATLAS/ISwR.out

8 	laGP 	1.5-5 	ATLAS 	https://www.stats.ox.ac.uk/pub/bdr/Rblas/ATLAS/laGP.out

48 	AGread 	1.1.1 	clang-UBSAN 	https://www.stats.ox.ac.uk/pub/bdr/memtests/clang-UBSAN/AGread

49 	BASIX 	1.1 	clang-UBSAN 	https://www.stats.ox.ac.uk/pub/bdr/memtests/clang-UBSAN/BASIX

51 	BayesSenMC 	0.1.4 	clang-UBSAN 	https://www.stats.ox.ac.uk/pub/bdr/memtests/clang-UBSAN/BayesSenMC

53 	BoolNet 	2.1.5 	clang-UBSAN 	https://www.stats.ox.ac.uk/pub/bdr/memtests/clang-UBSAN/BoolNet

## Show counts of issues according to kind:
table(issues[, "kind"])

## 
##       ATLAS clang-UBSAN    donttest    gcc-ASAN   gcc-UBSAN       gcc11         LTO       M1mac         MKL 
##           2          39          51           1          16           3           1          50           5 
##        noLD       noOMP  noSuggests    OpenBLAS        rchk      rlibro    valgrind 
##           6           2          11           7          83           2           6

## Summarize CRAN check status for 10 randomly-selected packages
## (reusing the information already read in):
pos <- sample(seq_len(NROW(pdb)), 10L)
summarize_CRAN_check_status(pdb[pos, "Package"],
                            results, details, issues)

## Package: bvenn
## **************
## 
## Current CRAN status: ERROR: 0, NOTE: 6, OK: 8
## See: <https://CRAN.R-project.org/web/checks/check_results_bvenn.html>
## 
## Version: 0.1
## Check: Rd line widths, Result: NOTE
##   Rd file 'bvenn.Rd':
##     \examples lines wider than 100 characters:
##                        bvenn(list(Set1 = sample(letters, 16), Set2 = sample(letters, 3+ 3*j), Set3 = sample(letters, 7)), add = TRUE, fontsize ... [TRUNCATED]
##   
##   These lines will be truncated in the PDF manual.
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/bvenn-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/bvenn-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-clang/bvenn-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/bvenn-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-patched-linux-x86_64/bvenn-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-linux-x86_64/bvenn-00check.html>
## 
## Package: MCPModGeneral
## **********************
## 
## Current CRAN status: ERROR: 0, NOTE: 11, OK: 2
## See: <https://CRAN.R-project.org/web/checks/check_results_MCPModGeneral.html>
## 
## Version: 0.1-1
## Check: LazyData, Result: NOTE
##     'LazyData' is specified without a 'data' directory
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-clang/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-UL/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-TK/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-old/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-patched-linux-x86_64/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-linux-x86_64/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-arm64/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-x86_64/MCPModGeneral-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-windows-ix86+x86_64/MCPModGeneral-00check.html>
## 
## Version: 0.1-1
## Check: re-building of vignette outputs, Result: WARNING
##   Error(s) in re-building vignettes:
##     ...
##   --- re-building ‘MCPModGeneral-Vignette.Rmd’ using rmarkdown
##   ! You can't use `macro parameter character #' in math mode.
##   l.353 migraine\(ntrtmigraine#
##                                > dose painfree ntrt pfrat #> 1 0.0 13 133 0.09...
##   
##   Error: processing vignette ‘MCPModGeneral-Vignette.Rmd’ failed with diagnostics:
##   LaTeX failed to compile /home/hornik/tmp/R.check/r-devel-gcc/Work/PKGS/MCPModGeneral.Rcheck/vign_test/MCPModGeneral/vignettes/MCPModGeneral-Vignette.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See MCPModGeneral-Vignette.log for more info.
##   --- failed re-building ‘MCPModGeneral-Vignette.Rmd’
##   
##   SUMMARY: processing the following file failed:
##     ‘MCPModGeneral-Vignette.Rmd’
##   
##   Error: Vignette re-building failed.
##   Execution halted
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/MCPModGeneral-00check.html>
## 
## Package: BAwiR
## **************
## 
## Current CRAN status: ERROR: 0, NOTE: 0, OK: 14
## See: <https://CRAN.R-project.org/web/checks/check_results_BAwiR.html>
## 
## Package: MixtureInf
## *******************
## 
## Current CRAN status: ERROR: 0, NOTE: 0, OK: 14
## See: <https://CRAN.R-project.org/web/checks/check_results_MixtureInf.html>
## 
## Package: mstknnclust
## ********************
## 
## Current CRAN status: ERROR: 0, NOTE: 0, OK: 14
## See: <https://CRAN.R-project.org/web/checks/check_results_mstknnclust.html>
## 
## Package: ICAMS
## **************
## 
## Current CRAN status: ERROR: 0, NOTE: 3, OK: 11
## See: <https://CRAN.R-project.org/web/checks/check_results_ICAMS.html>
## 
## Version: 2.3.10
## Check: package dependencies, Result: NOTE
##   Package suggested but not available for checking: ‘BSgenome.Mmusculus.UCSC.mm10’
## See: <https://www.r-project.org/nosvn/R.check/r-release-macos-arm64/ICAMS-00check.html>
## 
## Version: 2.3.10
## Check: package dependencies, Result: NOTE
##   Packages suggested but not available for checking:
##     'BSgenome.Hsapiens.1000genomes.hs37d5',
##     'BSgenome.Hsapiens.UCSC.hg38', 'BSgenome.Mmusculus.UCSC.mm10'
## See: <https://www.r-project.org/nosvn/R.check/r-release-macos-x86_64/ICAMS-00check.html>
## 
## Version: 2.3.10
## Check: package dependencies, Result: NOTE
##   Packages suggested but not available for checking:
##     'BSgenome.Hsapiens.1000genomes.hs37d5',
##     'BSgenome.Mmusculus.UCSC.mm10'
## See: <https://www.r-project.org/nosvn/R.check/r-oldrel-macos-x86_64/ICAMS-00check.html>
## 
## Package: tramnet
## ****************
## 
## Current CRAN status: ERROR: 0, NOTE: 0, OK: 12
## See: <https://CRAN.R-project.org/web/checks/check_results_tramnet.html>
## 
## Version: 0.0-6
## Check: re-building of vignette outputs, Result: WARNING
##   Error(s) in re-building vignettes:
##     ...
##   --- re-building ‘tramnet.Rnw’ using knitr
##   Loading required package: tram
##   Loading required package: mlt
##   Loading required package: basefun
##   Loading required package: variables
##   Loading required package: CVXR
##   
##   Attaching package: 'CVXR'
##   
##   The following object is masked from 'package:stats':
##   
##       power
##   
##   Loading required package: mlrMBO
##   Loading required package: mlr
##   Loading required package: ParamHelpers
##   Warning message: 'mlr' is in 'maintenance-only' mode since July
##   2019. Future development will only happen in 'mlr3'
##   (<https://mlr3.mlr-org.com>). Due to the focus on 'mlr3' there
##   might be uncaught bugs meanwhile in {mlr} - please consider
##   switching.
##   Loading required package: smoof
##   Loading required package: checkmate
##   Loading required package: survival
##   Welcome to penalized. For extended examples, see vignette("penalized").
##   Loading required package: Matrix
##   Loaded glmnet 4.1-3
##   
##   Attaching package: 'coin'
##   
##   The following object is masked from 'package:CVXR':
##   
##       size
##   
##   The following object is masked from 'package:variables':
##   
##       support
##   
##   
##   warning: solve(): system is singular (rcond: 1.91008e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   Error: processing vignette ‘tramnet.Rnw’ failed with diagnostics:
##   Running ‘texi2dvi’ on ‘tramnet.tex’ failed.
##   LaTeX errors:
##   ! Missing $ inserted.
##   <inserted text> 
##                   $
##   l.772 x <- model.matrix(~ 0 + x, ...)x_
##                                          scaled <- scale(x)mt <- tramnet(model...
##   ! Emergency stop.
##   <inserted text> 
##                   $
##   l.772 x <- model.matrix(~ 0 + x, ...)x_
##                                          scaled <- scale(x)mt <- tramnet(model...
##   !  ==> Fatal error occurred, no output PDF file produced!
##   --- failed re-building ‘tramnet.Rnw’
##   
##   SUMMARY: processing the following file failed:
##     ‘tramnet.Rnw’
##   
##   Error: Vignette re-building failed.
##   Execution halted
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/tramnet-00check.html>
## 
## Version: 0.0-6
## Check: re-building of vignette outputs, Result: WARNING
##   Error(s) in re-building vignettes:
##   --- re-building 'tramnet.Rnw' using knitr
##   Loading required package: tram
##   Loading required package: mlt
##   Loading required package: basefun
##   Loading required package: variables
##   Loading required package: CVXR
##   
##   Attaching package: 'CVXR'
##   
##   The following object is masked from 'package:stats':
##   
##       power
##   
##   Loading required package: mlrMBO
##   Loading required package: mlr
##   Loading required package: ParamHelpers
##   Warning message: 'mlr' is in 'maintenance-only' mode since July
##   2019. Future development will only happen in 'mlr3'
##   (<https://mlr3.mlr-org.com>). Due to the focus on 'mlr3' there
##   might be uncaught bugs meanwhile in {mlr} - please consider
##   switching.
##   Loading required package: smoof
##   Loading required package: checkmate
##   Loading required package: survival
##   Welcome to penalized. For extended examples, see vignette("penalized").
##   Loading required package: Matrix
##   Loaded glmnet 4.1-3
##   
##   Attaching package: 'coin'
##   
##   The following object is masked from 'package:CVXR':
##   
##       size
##   
##   The following object is masked from 'package:variables':
##   
##       support
##   
##   
##   warning: solve(): system is singular (rcond: 1.91008e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   
##   warning: solve(): system is singular; attempting approx solution
##   
##   warning: solve(): system is singular (rcond: 1.92453e-17); attempting approx solution
##   Warning in system(paste(shQuote(texi2dvi), if (quiet) "--quiet" else "",  :
##     running command '"C:\PROGRA~1\MiKTeX\miktex\bin\x64\texify.exe" --quiet --pdf "tramnet.tex" --max-iterations=20 -I "C:/r_packages/rinst/share/texmf/tex/latex" -I "C:/r_packages/rinst/share/texmf/bibtex/bst"' had status 1
##   Error: processing vignette 'tramnet.Rnw' failed with diagnostics:
##   running 'texi2dvi' on 'tramnet.tex' failed
##   
##   LaTeX errors:
##   ! LaTeX Error: File `tabu.sty' not found.
##   
##   Type X to quit or <RETURN> to proceed,
##   or enter new name. (Default extension: sty)
##   
##   
##   --- failed re-building 'tramnet.Rnw'
##   
##   SUMMARY: processing the following file failed:
##     'tramnet.Rnw'
##   
##   Error: Vignette re-building failed.
##   Execution halted
## See: <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-TK/tramnet-00check.html>
## 
## Package: intsurvbin
## *******************
## 
## Current CRAN status: ERROR: 0, NOTE: 13, OK: 1
## See: <https://CRAN.R-project.org/web/checks/check_results_intsurvbin.html>
## 
## Version: 0.0.4
## Check: LazyData, Result: NOTE
##     'LazyData' is specified without a 'data' directory
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-clang/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-UL/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-TK/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-old/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-patched-linux-x86_64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-linux-x86_64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-arm64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-x86_64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-windows-ix86+x86_64/intsurvbin-00check.html>
## 
## Version: 0.0.4
## Check: dependencies in R code, Result: NOTE
##   Namespace in Imports field not imported from: ‘mgcv’
##     All declared Imports should be used.
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-clang/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-TK/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-arm64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-macos-x86_64/intsurvbin-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-oldrel-macos-x86_64/intsurvbin-00check.html>
## 
## Package: SILM
## *************
## 
## Current CRAN status: ERROR: 0, NOTE: 0, OK: 14
## See: <https://CRAN.R-project.org/web/checks/check_results_SILM.html>
## 
## Package: crossrun
## *****************
## 
## Current CRAN status: ERROR: 0, NOTE: 8, OK: 6
## See: <https://CRAN.R-project.org/web/checks/check_results_crossrun.html>
## 
## Version: 0.1.0
## Check: package vignettes, Result: NOTE
##   Package vignette with placeholder title 'Vignette Title':
##     'vignettecrossrun.Rmd'
## See: <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-gcc/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-new-UL/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-devel-windows-x86_64-old/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-patched-linux-x86_64/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-linux-x86_64/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-release-windows-ix86+x86_64/crossrun-00check.html>,
##      <https://www.r-project.org/nosvn/R.check/r-oldrel-windows-ix86+x86_64/crossrun-00check.html>

## End(No test)

[Package tools version 4.1.2 Index]