Rd2HTML {tools}	R Documentation

Work with an Rd object

Description

These experimental functions take the output of the parse_Rd function and check it or produce a help page from it. Their interfaces (and existence!) are subject to change.

Usage

Rd2HTML(Rd, out = "", package = "", defines = .Platform$OS.type,
        encoding = "unknown", Links = NULL, CHM = FALSE)

Rd2txt(Rd, out = "", package = "", defines = .Platform$OS.type)

Rd2latex(Rd, out = "", defines = .Platform$OS.type, encoding = "unknown")

Rd2ex(Rd, out = "", defines = .Platform$OS.type, encoding = "unknown")

checkRd(Rd, defines = .Platform$OS.type, encoding = "unknown",
        unknownOK = FALSE, listOK = TRUE)

findHTMLlinks(pkgDir = "", lib.loc = NULL)

Arguments

Rd 	 the Rd object.

out 	 a filename or connection object to which to write the output.

package 	 the package to list in the output.

defines 	 string(s) to use in "#ifdef" tests.

encoding 	 encoding to use if the file does not specify one.

Links 	 NULL or a named (by topics) character vector of links, as returned by findHTMLlinks.

CHM 	 logical: prepare HTML suitable for Compiled HTML?

unknownOK 	 unrecognized macros are treated as errors if FALSE, otherwise warnings.

listOK 	 unnecessary non-empty braces (e.g. around text, not as an argument) are treated as errors if FALSE, otherwise warnings.

pkgDir 	 The top-level directory of an installed package.

lib.loc 	 character vector describing the location of R library trees to scan: the default indicates .libPaths().

Details

The Rd2HTML converter is still in development, and currently has the following known deficiencies:

"\enc" is not supported.
Error messages are not as informative as they could be. Often using parse_Rd(..., verbose = TRUE) will give a better idea of what is wrong.
Formatting is not always optimal.

Rd2ex extracts the examples in the format used by example and R utilities. it reproduces the previous version up to

The parser does not interpret markup in R comments: this used to be done.
Small improvements in layout.

findHTMLlinks uses in decreasing priority

The package in pkgDir: this is used when converting HTML help for that package.
The packages that are loaded in a vanilla R session.
Other packages found in the library trees specified by lib.loc in the order of the trees and alphabetically within a library tree.

Value

Rd2* are executed mainly for the side effect of writing the converted help page. Its value is the name of that page.
checkRd returns TRUE if the file has no errors; it generates an R error otherwise.
findHTMLlinks returns a named character vector of relative file paths, in the same package or in a package in the same library tree.

Author(s)

Duncan Murdoch, Brian Ripley

References

http://developer.r-project.org/parseRd.pdf

See Also

parse_Rd

Examples

toolsRd <- Rd_db("tools")
con <- textConnection(toolsRd[[grep("Rd2HTML.Rd", names(toolsRd))]], "rt")
outfile <- paste(tempfile(), ".html", sep="")
browseURL(Rd2HTML(con, outfile, package="tools"))
close(con)
 
con <- textConnection(toolsRd[["Rd2HTML.Rd"]], "rt")
checkRd(con) # A stricter test than Rd2HTML uses
close(con)

[Package tools version 2.9.2 Index]