Rdutils {tools}	R Documentation

Rd Utilities

Description

Utilities for computing on the information in Rd objects.

Usage

Rd_db(package, dir, lib.loc = NULL, stages = "build")

Arguments

package
a character string naming an installed package.

dir
a character string specifying the path to a package's root source directory. This should contain the subdirectory ‘man’ with R documentation sources (in Rd format). Only used if package is not given.

lib.loc
a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. The specified library trees are used to search for package.

stages
if dir is specified and the database is being built from source, which stages of \Sexpr processing should be processed?

Details

Rd_db builds a simple database of all Rd objects in a package, as a list of the results of running parse_Rd on the Rd source files in the package and processing platform conditionals and some \Sexpr macros.

See Also

parse_Rd

Examples

## No test: 
## Build the Rd db for the (installed) base package.
db <- Rd_db("base")

## Keyword metadata per Rd object.
keywords <- lapply(db, tools:::.Rd_get_metadata, "keyword")
## Tabulate the keyword entries.
kw_table <- sort(table(unlist(keywords)))
## The 5 most frequent ones:
rev(kw_table)[1 : 5]

## 
##   utilities programming        file       array       manip 
##          85          75          45          42          35

## The "most informative" ones:
kw_table[kw_table == 1]

## 
## complex  device   files   hplot package 
##       1       1       1       1       1

## Concept metadata per Rd file.
concepts <- lapply(db, tools:::.Rd_get_metadata, "concept")
## How many files already have \concept metadata?
sum(sapply(concepts, length) > 0)

## [1] 35

## How many concept entries altogether?
length(unlist(concepts))

## [1] 74

## End(No test)

[Package tools version 4.1.2 Index]