buildVignettes {tools}	R Documentation

List and Build Package Vignettes

Description

Run Sweave and texi2dvi on all vignettes of a package.

Usage

buildVignettes(package, dir, lib.loc = NULL, quiet = TRUE, clean = TRUE)
pkgVignettes(package, dir, lib.loc = NULL)

Arguments

package 	 a character string naming an installed package. If given, Sweave files are searched in subdirectory ‘doc’.

dir 	 a character string specifying the path to a package's root source directory. This subdirectory ‘inst/doc’ is searched for Sweave files.

lib.loc 	 a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. The specified library trees are used to search for package.

quiet 	 logical. Run Sweave and texi2dvi in quiet mode.

clean 	 Remove all files generated by the build, even if there were copies there before.

Value

buildVignettes is called for its side effect of creating the PDF versions of all vignettes.
pkgVignettes returns an object of class "pkgVignettes".

[Package tools version 2.9.2 Index]