buildVignettes {tools}	R Documentation

List and Build Package Vignettes

Description

Run Sweave (or other custom weave function) and texi2dvi on all vignettes of a package.

Usage


buildVignettes(package, dir, lib.loc = NULL, quiet = TRUE,
               clean = TRUE, tangle = FALSE, ser_elibs = NULL)

pkgVignettes(package, dir, subdirs = NULL, lib.loc = NULL,
	     output = FALSE, source = FALSE, check = FALSE)

Arguments

package
a character string naming an installed package. If given, vignette source files are by default looked for in subdirectory ‘doc’.

dir
a character string specifying the path to a package's root source directory. This subdirectory ‘vignettes’ (or if it does not exist ‘inst/doc’) is searched for vignette source files.

lib.loc
a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. The specified library trees are used to search for package.

quiet
logical. Weave and run texi2pdf in quiet mode.

clean
Remove all files generated by the build, even if there were copies there before.

tangle
logical. Do tangling as well as weaving.

ser_elibs
For use from R CMD check.

subdirs
a character vector of subdirectories of dir in which to look for vignettes. The first which exists is used. Defaults to "doc" if package is supplied, otherwise "vignettes".

output
logical indicating if the output filenames for each vignette should be returned (in component outputs).

source
logical indicating if the tangled output filenames for each vignette should be returned (in component sources).

check
logical. If TRUE, check whether all files that have vignette-like filenames have an identifiable vignette engine. This may be a false positive if a file is not a vignette but has a filename matching a pattern defined by one of the vignette engines.

Details

buildVignettes is used by R CMD build and R CMD check to (re-)build vignette PDFs from their sources.

As from R 3.4.1, both of these functions ignore files that are listed in the ‘.Rbuildignore’ file in dir.

Value

buildVignettes is called for its side effect of creating the PDF versions of all vignettes, and if tangle = TRUE, extracting the R code.

pkgVignettes returns an object of class "pkgVignettes" if a vignette directory is found, otherwise NULL.

Examples


gVigns <- pkgVignettes("grid")
str(gVigns)

[Package tools version 3.6.1 Index]