charsets {tools}	R Documentation

Conversion Tables between Character Sets

Description

charset_to_Unicode is a matrix of Unicode code points with columns for the common 8-bit encodings.

Adobe_glyphs is a data frame which gives Adobe glyph names for Unicode code points. It has two character columns, "adobe" and "unicode" (a 4-digit hex representation).

Usage

charset_to_Unicode

Adobe_glyphs

Details

charset_to_Unicode is an integer matrix of class c("noquote", "hexmode") so prints in hexadecimal. The mappings are those used by libiconv: there are differences in the way quotes and minus/hyphen are mapped between sources (and the postscript encoding files use a different mapping).

Adobe_glyphs includes all the Adobe glyph names which correspond to single Unicode characters. It is sorted by Unicode code point and within a point alphabetically on the glyph (there can be more than one name for a Unicode code point). The data are in the file ‘R_HOME/share/encodings/Adobe_glyphlist’.

Examples

## find Adobe names for ISOLatin2 chars.
latin2 <- charset_to_Unicode[, "ISOLatin2"]
aUnicode <- as.hexmode(paste0("0x", Adobe_glyphs$unicode))
keep <- aUnicode %in% latin2
aUnicode <- aUnicode[keep]
aAdobe <- Adobe_glyphs[keep, 1]
## first match
aLatin2 <- aAdobe[match(latin2, aUnicode)]
## all matches
bLatin2 <- lapply(1:256, function(x) aAdobe[aUnicode == latin2[x]])
format(bLatin2, justify = "none")

##   [1] ""                          "controlSTX"                "controlSOT"               
##   [4] "controlETX"                "controlEOT"                "controlENQ"               
##   [7] "controlACK"                "controlBEL"                "controlBS"                
##  [10] "controlHT"                 "controlLF"                 "controlVT"                
##  [13] "controlFF"                 "controlCR"                 "controlSO"                
##  [16] "controlSI"                 "controlDLE"                "controlDC1"               
##  [19] "controlDC2"                "controlDC3"                "controlDC4"               
##  [22] "controlNAK"                "controlSYN"                "controlETB"               
##  [25] "controlCAN"                "controlEM"                 "controlSUB"               
##  [28] "controlESC"                "controlFS"                 "controlGS"                
##  [31] "controlRS"                 "controlUS"                 "space, spacehackarabic"   
##  [34] "exclam"                    "quotedbl"                  "numbersign"               
##  [37] "dollar"                    "percent"                   "ampersand"                
##  [40] "quotesingle"               "parenleft"                 "parenright"               
##  [43] "asterisk"                  "plus"                      "comma"                    
##  [46] "hyphen"                    "period"                    "slash"                    
##  [49] "zero"                      "one"                       "two"                      
##  [52] "three"                     "four"                      "five"                     
##  [55] "six"                       "seven"                     "eight"                    
##  [58] "nine"                      "colon"                     "semicolon"                
##  [61] "less"                      "equal"                     "greater"                  
##  [64] "question"                  "at"                        "A"                        
##  [67] "B"                         "C"                         "D"                        
##  [70] "E"                         "F"                         "G"                        
##  [73] "H"                         "I"                         "J"                        
##  [76] "K"                         "L"                         "M"                        
##  [79] "N"                         "O"                         "P"                        
##  [82] "Q"                         "R"                         "S"                        
##  [85] "T"                         "U"                         "V"                        
##  [88] "W"                         "X"                         "Y"                        
##  [91] "Z"                         "bracketleft"               "backslash"                
##  [94] "bracketright"              "asciicircum"               "underscore"               
##  [97] "grave"                     "a"                         "b"                        
## [100] "c"                         "d"                         "e"                        
## [103] "f"                         "g"                         "h"                        
## [106] "i"                         "j"                         "k"                        
## [109] "l"                         "m"                         "n"                        
## [112] "o"                         "p"                         "q"                        
## [115] "r"                         "s"                         "t"                        
## [118] "u"                         "v"                         "w"                        
## [121] "x"                         "y"                         "z"                        
## [124] "braceleft"                 "bar, verticalbar"          "braceright"               
## [127] "asciitilde"                "controlDEL"                ""                         
## [130] ""                          ""                          ""                         
## [133] ""                          ""                          ""                         
## [136] ""                          ""                          ""                         
## [139] ""                          ""                          ""                         
## [142] ""                          ""                          ""                         
## [145] ""                          ""                          ""                         
## [148] ""                          ""                          ""                         
## [151] ""                          ""                          ""                         
## [154] ""                          ""                          ""                         
## [157] ""                          ""                          ""                         
## [160] ""                          "nbspace, nonbreakingspace" "Aogonek"                  
## [163] "breve"                     "Lslash"                    "currency"                 
## [166] "Lcaron"                    "Sacute"                    "section"                  
## [169] "dieresis"                  "Scaron"                    "Scedilla"                 
## [172] "Tcaron"                    "Zacute"                    "sfthyphen, softhyphen"    
## [175] "Zcaron"                    "Zdot, Zdotaccent"          "degree"                   
## [178] "aogonek"                   "ogonek"                    "lslash"                   
## [181] "acute"                     "lcaron"                    "sacute"                   
## [184] "caron"                     "cedilla"                   "scaron"                   
## [187] "scedilla"                  "tcaron"                    "zacute"                   
## [190] "hungarumlaut"              "zcaron"                    "zdot, zdotaccent"         
## [193] "Racute"                    "Aacute"                    "Acircumflex"              
## [196] "Abreve"                    "Adieresis"                 "Lacute"                   
## [199] "Cacute"                    "Ccedilla"                  "Ccaron"                   
## [202] "Eacute"                    "Eogonek"                   "Edieresis"                
## [205] "Ecaron"                    "Iacute"                    "Icircumflex"              
## [208] "Dcaron"                    "Dcroat, Dslash"            "Nacute"                   
## [211] "Ncaron"                    "Oacute"                    "Ocircumflex"              
## [214] "Odblacute, Ohungarumlaut"  "Odieresis"                 "multiply"                 
## [217] "Rcaron"                    "Uring"                     "Uacute"                   
## [220] "Udblacute, Uhungarumlaut"  "Udieresis"                 "Yacute"                   
## [223] "Tcedilla, Tcommaaccent"    "germandbls"                "racute"                   
## [226] "aacute"                    "acircumflex"               "abreve"                   
## [229] "adieresis"                 "lacute"                    "cacute"                   
## [232] "ccedilla"                  "ccaron"                    "eacute"                   
## [235] "eogonek"                   "edieresis"                 "ecaron"                   
## [238] "iacute"                    "icircumflex"               "dcaron"                   
## [241] "dcroat, dmacron"           "nacute"                    "ncaron"                   
## [244] "oacute"                    "ocircumflex"               "odblacute, ohungarumlaut" 
## [247] "odieresis"                 "divide"                    "rcaron"                   
## [250] "uring"                     "uacute"                    "udblacute, uhungarumlaut" 
## [253] "udieresis"                 "yacute"                    "tcedilla, tcommaaccent"   
## [256] "dotaccent"

[Package tools version 4.1.2 Index]