checkMD5sums {tools}	R Documentation

Check and Create MD5 Checksum Files

Description

checkMD5sums checks the files against a file ‘MD5’.

Usage


checkMD5sums(package, dir)

Arguments

package
the name of an installed package

dir
the path to the top-level directory of an installed package.

Details

The file ‘MD5’ which is created is in a format which can be checked by md5sum -c MD5 if a suitable command-line version of md5sum is available. (For Windows, one is supplied in the bundle at https://cran.r-project.org/bin/windows/Rtools.)

If dir is missing, an installed package of name package is searched for.

The private function tools:::.installMD5sums is used to create MD5 files in the Windows build.

Value

checkMD5sums returns a logical, NA if there is no ‘MD5’ file to be checked.

See Also

md5sum

[Package tools version 3.6.1 Index]