checkVignettes {tools}	R Documentation

Check Package Vignettes

Description

Check all Sweave files of a package by running Sweave and/or Stangle on them. All R source code files found after the tangling step are sourceed to check whether all code can be executed without errors.

Usage

checkVignettes(package, dir, lib.loc = NULL,
               tangle = TRUE, weave = TRUE, latex = FALSE,
               workdir = c("tmp", "src", "cur"),
               keepfiles = FALSE)

Arguments

package 	 a character string naming an installed package. If given, Sweave files are searched in subdirectory ‘doc’.

dir 	 a character string specifying the path to a package's root source directory. This subdirectory ‘inst/doc’ is searched for Sweave files.

lib.loc 	 a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known. The specified library trees are used to search for package.

tangle 	 Perform a tangle and source the extracted code?

weave 	 Perform a weave?

latex 	 logical: if tangle, weave and latex are TRUE and there is no ‘Makefile’ in the vignettes directory, run the tangled files through pdflatex.

workdir 	 Directory used as working directory while checking the vignettes. If "tmp" then a temporary directory is created, this is the default. If "src" then the directory containing the vignettes itself is used, if "cur" then the current working directory of R is used.

keepfiles 	 Delete file in temporary directory? This option is ignored when workdir != "tmp".

Value

An object of class "checkVignettes" which is a list with the error messages found during the tangle and weave steps. There is a print method for nicely displaying the information contained in such objects.

[Package tools version 2.9.2 Index]