delimMatch {tools}	R Documentation

Delimited Pattern Matching

Description

Match delimited substrings in a character vector, with proper nesting.

Usage


delimMatch(x, delim = c("{", "}"), syntax = "Rd")

Arguments

x
a character vector.

delim
a character vector of length 2 giving the start and end delimiters. Future versions might allow for arbitrary regular expressions.

syntax
currently, always the string "Rd" indicating Rd syntax (i.e., % starts a comment extending till the end of the line, and \ escapes). Future versions might know about other syntax, perhaps via ‘syntax tables’ allowing to flexibly specify comment, escape, and quote characters.

Value

An integer vector of the same length as x giving the starting position (in characters) of the first match, or -1 if there is none, with attribute "match.length" giving the length (in characters) of the matched text (or -1 for no match).

See Also

regexpr for ‘simple’ pattern matching.

Examples


x <- c("\\value{foo}", "function(bar)")
delimMatch(x)
delimMatch(x, c("(", ")"))

[Package tools version 3.6.1 Index]