dependsOnPkgs {tools}	R Documentation

Find Reverse Dependencies

Description

Find ‘reverse’ dependencies of packages, that is those packages which depend on this one, and (optionally) so on recursively.

Usage

dependsOnPkgs(pkgs, dependencies = c("Depends", "Imports"),
              recursive = TRUE, lib.loc = NULL,
              installed = installed.packages(lib.loc, fields = "Enhances"))

Arguments

pkgs 	 a character vector of package names.

dependencies 	 a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "Suggests", "Enhances").

recursive 	 logical: should reverse dependencies of reverse dependencies (and so on) be included?

lib.loc 	 a character vector of R library trees, or NULL for all known trees (see .libPaths).

installed 	 a result of calling installed.packages.

Value

A character vector of package names, which does not include any from pkgs.

Examples

## there are few dependencies in a vanilla R installation
dependsOnPkgs("lattice")

[Package tools version 2.9.2 Index]