getVignetteInfo {tools}	R Documentation

Get Information on Installed Vignettes

Description

This function gets information on installed vignettes.

Usage

getVignetteInfo(package = NULL, lib.loc = NULL, all = TRUE)

Arguments

package
Which package to look in, or NULL for all packages.

lib.loc
Which library to look in.

all
Whether to search all installed packages, or just attached packages.

Value

A matrix with columns

Package
the name of the package

Dir
the directory where the package is installed

Topic
the name of the vignette

File
the base filename of the source of the vignette

Title
the title of the vignette

R
the tangled R source from the vignette

PDF
the PDF or HTML file for display

Note

The last column of the result is named PDF for historical reasons, but it may contain a filename of a PDF or HTML document.

See Also

pkgVignettes is a similar function that can work on an uninstalled package.

Examples

## No test: 
getVignetteInfo("grid")

##       Package Dir                                                      Topic         File             
##  [1,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "moveline"    "moveline.Rnw"   
##  [2,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "displaylist" "displaylist.Rnw"
##  [3,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "interactive" "interactive.Rnw"
##  [4,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "frame"       "frame.Rnw"      
##  [5,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "grid"        "grid.Rnw"       
##  [6,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "locndimn"    "locndimn.Rnw"   
##  [7,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "sharing"     "sharing.Rnw"    
##  [8,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "nonfinite"   "nonfinite.Rnw"  
##  [9,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "saveload"    "saveload.Rnw"   
## [10,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "rotated"     "rotated.Rnw"    
## [11,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "grobs"       "grobs.Rnw"      
## [12,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "viewports"   "viewports.Rnw"  
## [13,] "grid"  "/Library/Frameworks/R.framework/Resources/library/grid" "plotexample" "plotexample.Rnw"
##       Title                                     R               PDF              
##  [1,] "Demonstrating move-to and line-to"       "moveline.R"    "moveline.pdf"   
##  [2,] "Display Lists in grid"                   "displaylist.R" "displaylist.pdf"
##  [3,] "Editing grid Graphics"                   "interactive.R" "interactive.pdf"
##  [4,] "Frames and packing grobs"                "frame.R"       "frame.pdf"      
##  [5,] "Introduction to grid"                    "grid.R"        "grid.pdf"       
##  [6,] "Locations versus Dimensions"             "locndimn.R"    "locndimn.pdf"   
##  [7,] "Modifying multiple grobs simultaneously" "sharing.R"     "sharing.pdf"    
##  [8,] "Non-finite values"                       "nonfinite.R"   "nonfinite.pdf"  
##  [9,] "Persistent representations"              "saveload.R"    "saveload.pdf"   
## [10,] "Rotated Viewports"                       "rotated.R"     "rotated.pdf"    
## [11,] "Working with grid grobs"                 "grobs.R"       "grobs.pdf"      
## [12,] "Working with viewports"                  "viewports.R"   "viewports.pdf"  
## [13,] "Writing grid Code"                       "plotexample.R" "plotexample.pdf"

## End(No test)

[Package tools version 4.1.2 Index]