loadRdMacros {tools}	R Documentation

Load User-defined Rd Help System Macros

Description

Loads macros from an ‘.Rd’ file, or from several ‘.Rd’ files contained in a package.

Usage

loadRdMacros(file, macros = TRUE)
loadPkgRdMacros(pkgdir, macros)

Arguments

file
A file in Rd format containing macro definitions.

macros
TRUE or a previous set of macro definitions, in the format expected by the parse_Rd macros argument.

pkgdir
The base directory of a source package or an installed package.

Details

The files parsed by this function should contain only macro definitions; a warning will be issued if anything else other than comments or white space is found.

The macros argument may be a filename of a base set of macros, or the result of a previous call to loadRdMacros or loadPkgRdMacros in the same session. These results should be assumed to be valid only within the current session.

The loadPkgRdMacros function first looks for an "RdMacros" entry in the package ‘DESCRIPTION’ file. If present, it should contain a comma-separated list of other package names; their macros will be loaded before those of the current package. It will then look in the current package for ‘.Rd’ files in the ‘man/macros’ or ‘help/macros’ subdirectories, and load those.

Value

These functions each return an environment containing objects with the names of the newly defined macros from the last file processed. The parent environment will be macros from the previous file, and so on. The first file processed will have emptyenv() as its parent.

Author(s)

Duncan Murdoch

References

See the ‘Writing R Extensions’ manual for the syntax of Rd files, or https://developer.r-project.org/parseRd.pdf for a technical discussion.

See Also

parse_Rd

Examples

## No test: 
f <- tempfile()
writeLines(paste0("\\newcommand{\\logo}{\\if{html}{\\figure{Rlogo.svg}{options: width=100}",
                  "\\if{latex}{\\figure{Rlogo.pdf}{options: width=0.5in}}}"),
        f)
m <- loadRdMacros(f)

## Warning: /var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T//Rtmpmdh4o1/fileb4c153a5c0ea:2: unexpected END_OF_INPUT '\if{html}{\figure{Rlogo.svg}{options: width=100}\if{latex}{\figure{Rlogo.pdf}{options: width=0.5in}}}
## '

ls(m)

## [1] "\\logo"

ls(parent.env(m))

##  [1] "\\acronym"      "\\alias"        "\\arguments"    "\\author"       "\\bold"         "\\cite"        
##  [7] "\\code"         "\\command"      "\\concept"      "\\cr"           "\\deqn"         "\\describe"    
## [13] "\\description"  "\\details"      "\\dfn"          "\\docType"      "\\dontrun"      "\\dontshow"    
## [19] "\\donttest"     "\\dots"         "\\dQuote"       "\\email"        "\\emph"         "\\enc"         
## [25] "\\encoding"     "\\enumerate"    "\\env"          "\\eqn"          "\\examples"     "\\figure"      
## [31] "\\file"         "\\format"       "\\href"         "\\if"           "\\ifelse"       "\\item"        
## [37] "\\itemize"      "\\kbd"          "\\keyword"      "\\ldots"        "\\link"         "\\linkS4class" 
## [43] "\\method"       "\\name"         "\\newcommand"   "\\note"         "\\option"       "\\out"         
## [49] "\\pkg"          "\\preformatted" "\\R"            "\\RdOpts"       "\\Rdversion"    "\\references"  
## [55] "\\renewcommand" "\\S3method"     "\\S4method"     "\\samp"         "\\section"      "\\seealso"     
## [61] "\\Sexpr"        "\\source"       "\\special"      "\\sQuote"       "\\strong"       "\\subsection"  
## [67] "\\synopsis"     "\\tab"          "\\tabular"      "\\testonly"     "\\title"        "\\url"         
## [73] "\\usage"        "\\value"        "\\var"          "\\verb"         "#endif"         "#ifdef"        
## [79] "#ifndef"

ls(parent.env(parent.env(m)))

## character(0)

## End(No test)

[Package tools version 4.1.2 Index]