makeLazyLoading {tools}	R Documentation

Lazy Loading of Packages

Description

Tools for lazy loading of packages from a database.

Usage

makeLazyLoading(package, lib.loc = NULL, compress = TRUE,
                keep.source = getOption("keep.source.pkgs"))

Arguments

package 	 package name string

lib.loc 	 library trees, as in library

keep.source 	 logical; should sources be kept when saving from source

compress 	 logical; whether to compress entries on the database.

Details

A tool to set up packages for lazy loading from a database. For packages other than base you can use makeLazyLoading(package) to convert them to use lazy loading.

Author(s)

Luke Tierney and Brian Ripley

Examples

## set up package "splines" for lazy loading -- already done
## Not run: 
tools:::makeLazyLoading("splines")
## End(Not run)

[Package tools version 2.9.2 Index]