makevars {tools}	R Documentation

User and Site Compilation Variables

Description

Determine the location of the user and site specific ‘Makevars’ files for customizing package compilation.

Usage

makevars_user()
makevars_site()

Details

Package maintainers can use these functions to employ user and site specific compilation settings also for compilations not using R's mechanisms (in particular, custom compilations in subdirectories of ‘src’), e.g., by adding configure code calling R with cat(tools::makevars_user()) or cat(tools::makevars_site()), and if non-empty passing this with -f to custom Make invocations.

Value

A character string with the path to the user or site specific ‘Makevars’ file, or an empty character vector if there is no such file.

See Also

Section ‘Customizing package compilation’ in the ‘R Installation and Administration’ manual.

Examples

## No test: 
makevars_user()

## character(0)

makevars_site()

## character(0)

## End(No test)
## Don't show: 
checkMV <- function(r)
  stopifnot(is.character(r),
            length(r) == 0 || (length(r) == 1 && file.exists(r)))
checkMV(makevars_user())
checkMV(makevars_site())
## End(Don't show)

[Package tools version 4.1.2 Index]