md5sum {tools}	R Documentation

Compute MD5 Checksums

Description

Compute the 32-byte MD5 checksums of one or more files.

Usage

md5sum(files)

Arguments

files 	 character. The paths of file(s) to be check-summed.

Value

A character vector of the same length as files, with names equal to files. The elements will be NA for non-existent or unreadable files, otherwise a 32-character string of hexadecimal digits.
On Windows all files are read in binary mode (as the md5sum utilities there do): on other OSes the files are read in the default way.

See Also

checkMD5sums

Examples

md5sum(dir(R.home(), pattern="^COPY", full.names=TRUE))

[Package tools version 2.9.2 Index]