package.dependencies {tools}	R Documentation

Check Package Dependencies

Description

Parses and checks the dependencies of a package against the currently installed version of R (and other packages).

Usage

package.dependencies(x, check = FALSE,
                     depLevel = c("Depends", "Imports", "Suggests"))

Arguments

x 	 A matrix of package descriptions as returned by available.packages.

check 	 If TRUE, return logical vector of check results. If FALSE, return parsed list of dependencies.

depLevel 	 Whether to look for Depends or Suggests level dependencies.

Details

Currently we only check if the package conforms with the currently running version of R. In the future we might add checks for inter-package dependencies.

See Also

update.packages

[Package tools version 2.9.2 Index]