showNonASCII {tools}	R Documentation

Highlight non-ASCII characters

Description

This function prints elements of a character vector which contain non-ASCII bytes, printing such bytes as a escape like <fc>.

Usage

showNonASCII(x)

Arguments

x 	 a character vector.

Details

This was originally written to help detect non-portable text in files in packages.

It prints all lines of the files which contain non-ACII characters, preceded by the line number and with non-ASCII bytes highlighted via iconv(sub = "byte").

Examples

out <- c(
"fa\xE7ile test of showNonASCII():",
"\\details{",
"   This is a good line",
"   This has an \xfcmlaut in it.",
"   OK again.",
"}")
cat(out, file = "my.Rd", sep = "\n")

showNonASCII(readLines("my.Rd"))
unlink("my.Rd")

[Package tools version 2.9.2 Index]