testInstalledPackage {tools}	R Documentation

Test Installed Packages

Description

These functions allow an installed package to be tested, or all base and recommended packages.

Usage

testInstalledPackage(pkg, lib.loc = NULL, outDir = ".",
                     types = c("examples", "tests", "vignettes"))

testInstalledPackages(outDir = ".", errorsAreFatal = TRUE,
                      scope = c("both", "base", "recommended"),
                      types = c("examples", "tests", "vignettes"))

testInstalledBasic(scope = c("basic", "devel", "both"))

Arguments

pkg 	 name of an installed package.

lib.loc 	 library path(s) in which to look for the package. See library.

outDir 	 the directory into which to write the output files: this should already exist.

types 	 type(s) of tests to be done.

errorsAreFatal 	 logical: should testing terminate at the first error?

scope 	 Which set(s) should be tested?

Details

These tests depend on having the package example files installed (which is the default). If package-specific tests are found in a ‘tests’ directory they can be tested: these are not installed by default, but will be if R CMD INSTALL --install-tests was used. Finally, the R code in any vignettes can be extracted and tested.

Package tests are run in a ‘pkg-tests’ subdirectory of ‘outDir’, and leave their output there.

testInstalledBasic runs the basic tests, if installed. This should be run with LC_COLLATE=C set: the function tries to set this by it may not work on all OSes.

The package-specific tests for the base and recommended packages are an optional part of the install: for testing package tools you will also need the help page sources installed (also optional). Currently testing requires a Unix-like diff for a full report if more than simple differences are found.

Value

Invisbly 0L for success, 1L for failure.

[Package tools version 2.9.2 Index]