texi2dvi {tools}	R Documentation

Compile LaTeX Files

Description

Run latex and bibtex until all cross-references are resolved and create either a dvi or PDF file.

Usage

texi2dvi(file, pdf = FALSE, clean = FALSE, quiet = TRUE,
         texi2dvi = getOption("texi2dvi"))

Arguments

file 	 character. Name of TeX source file.

pdf 	 logical. If TRUE, a PDF file is produced insted of the default dvi file (texi2dvi command line option --pdf).

clean 	 logical. If TRUE, all auxiliary files are removed (texi2dvi command line option --clean). Does not work on some platforms.

quiet 	 logical. No output unless an error occurs.

texi2dvi 	 character (or NULL). Script or program used to compile a TeX file to dvi or PDF, respectively. If set to NULL, the ‘texi2dvi’ script in R's ‘bin’ directory is used (if it exists), otherwise it is assumed that texi2dvi is in the search path.

Details

Some TeX installations on Windows do not have ‘texi2dvi.exe’. If ‘texify.exe’ is present (e.g., part of MikTeX), then it can be used instead: set options(texi2dvi="texify.exe") or to the full path of the program.

Author(s)

Achim Zeileis

[Package tools version 2.1.0 Index]