texi2dvi {tools}	R Documentation

Compile LaTeX Files

Description

Run latex and bibtex until all cross-references are resolved and create either a dvi or PDF file.

Usage

texi2dvi(file, pdf = FALSE, clean = FALSE, quiet = TRUE,
         texi2dvi = getOption("texi2dvi"), texinputs = NULL)

Arguments

file 	 character. Name of LaTeX source file.

pdf 	 logical. If TRUE, a PDF file is produced instead of the default dvi file (texi2dvi command line option --pdf).

clean 	 logical. If TRUE, all auxiliary files are removed (texi2dvi command line option --clean). May not work on some platforms.

quiet 	 logical. No output unless an error occurs. Ignored if emulation (see the texi2dvi argument) is used.

texi2dvi 	 character (or NULL). Script or program used to compile a TeX file to dvi or PDF, respectively. The default (selected by "" or NULL) is to look for an executable on the search path and otherwise emulate the script with system calls.

texinputs 	 NULL or a character vector of paths to add to the LaTeX and bibtex input search paths.

Details

Despite the name, this is used in R to compile LaTeX files, specifically those generated from vignettes. It ensures that the ‘R_HOME/share/texmf’ directory is in the TEXINPUTS path, so R style files such as ‘Sweave’ and ‘Rd’ will be found. The search path used is first the existing TEXINPUTS setting (or the current directory if unset), then elements of texinputs, then ‘R_HOME/share/texmf’ and finally the default path. Analogous changes are made to BIBINPUTS and BSTINPUTSsettings.

MiKTeX has a texi2dvi executable but no other Windows TeX installation that we know of does, so emulation is used on e.g. TeXLive installations.

Author(s)

Achim Zeileis and R-core

[Package tools version 2.9.2 Index]