toRd {tools}	R Documentation

Generic Function to Convert Object to a Fragment of Rd Code

Description

Methods for this function render their associated classes as a fragment of Rd code, which can then be rendered into text, HTML, or LaTeX.

Usage


toRd(obj, ...)
## S3 method for class 'bibentry'
toRd(obj, style = NULL, ...)

Arguments

obj
The object to be rendered.

style
The style to be used in converting a bibentry object.

...
Additional arguments used by methods.

Details

See bibstyle for a discussion of styles. The default style = NULL value gives the default style.

Value

Returns a character vector containing a fragment of Rd code that could be parsed and rendered. The default method converts obj to mode character, then escapes any Rd markup within it. The bibentry method converts an object of that class to markup appropriate for use in a bibliography.

[Package tools version 3.6.1 Index]