tools-defunct {tools}	R Documentation

Defunct Objects in Package tools

Description

The functions or variables listed here are defunct, and may get removed as soon as of the next release.

See Also

tools-deprecated, Deprecated, Defunct

[Package tools version 2.9.2 Index]