tools-defunct {tools}	R Documentation

Defunct Objects in Package tools

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage


# Defunct in R 2.10.0, removed in R 3.0.0
Rd_parse(file, text = NULL)
# Defunct in R 3.1.0
readNEWS(file = file.path(R.home(), "NEWS"), trace = FALSE,
         chop = c("first", "1", "par1", "keepAll"))
checkNEWS(file = file.path(R.home(), "NEWS"))

See Also

tools-deprecated, Deprecated, Defunct

[Package tools version 3.6.1 Index]