tools-deprecated {tools}	R Documentation

Deprecated Objects in Package tools

Description

The functions or variables listed here are provided for compatibility with older versions of R only, and may be defunct as soon as of the next release.

See Also

Deprecated, Defunct

[Package tools version 2.9.2 Index]