tools-deprecated {tools}	R Documentation

Deprecated Objects in Package tools

Description

The functions or variables listed here are provided for compatibility with older versions of R only, and may be defunct as soon as of the next release.

Usage


package.dependencies(x, check = FALSE,
                     depLevel = c("Depends", "Imports", "Suggests"))

getDepList(depMtrx, instPkgs, recursive = TRUE, local = TRUE,
           reduce = TRUE, lib.loc = NULL)

pkgDepends(pkg, recursive = TRUE, local = TRUE, reduce = TRUE,
           lib.loc = NULL)

installFoundDepends(depPkgList, ...)

vignetteDepends(vignette, recursive = TRUE, reduce = TRUE,
                local = TRUE, lib.loc = NULL)

Arguments

x
A matrix of package descriptions as returned by available.packages.

check
If TRUE, return logical vector of check results. If FALSE, return parsed list of dependencies.

depLevel
Whether to look for Depends or Suggests level dependencies. Can be abbreviated.

depMtrx
a dependency matrix as from package.dependencies().

pkg
the name of the package

instPkgs
a matrix specifying all packages installed on the local system, as from installed.packages

recursive
whether or not to include indirect dependencies.

local
whether or not to search only locally

reduce
whether or not to collapse all sets of dependencies to a minimal value

lib.loc
what libraries to use when looking for installed packages. NULL indicates all library directories in the current .libPaths(). Note that lib.loc is not used in getDepList() and deprecated there.

depPkgList
A Found element from a pkgDependsList object

...
Arguments to pass on to install.packages

vignette
the path to the vignette source

See Also

Deprecated, Defunct

[Package tools version 3.6.1 Index]