vignetteDepends {tools}	R Documentation

Retrieve Dependency Information for a Vignette

Description

Given a vignette name, will create a DependsList object that reports information about the packages the vignette depends on.

Usage

vignetteDepends(vignette, recursive = TRUE, reduce = TRUE,
                local = TRUE, lib.loc = NULL)

Arguments

vignette 	 The path to the vignette source

recursive 	 Whether or not to include indirect dependencies

reduce 	 Whether or not to collapse all sets of dependencies to a minimal value

local 	 Whether or not to search only locally

lib.loc 	 What libraries to search in locally

Details

If recursive is TRUE, any package that is specified as a dependency will in turn have its dependencies included (and so on), these are known as indirect dependencies. If recursive is FALSE, only the dependencies directly named by the vignette will be used.

If local is TRUE, the system will only look at the user's local machine and not online to find dependencies.

If reduce is TRUE, the system will collapse the fields in the DependsList object such that a minimal set of dependencies are specified (for instance if there was foo, foo (>= 1.0.0), foo (>= 1.3.0, it would only return foo (>= 1.3.0)).

Value

An object of class "DependsList".

Author(s)

Jeff Gentry

See Also

pkgDepends

Examples

gridEx <- system.file("doc", "grid.Snw", package = "grid")
vignetteDepends(gridEx)

[Package tools version 2.9.2 Index]